requirements.txt
supabase>=2.8.0
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
zai-sdk==0.1.0
httpx>=0.28.0
mangum>=0.17.0
pydantic>=2.0.0

